Refactor translations into separate modules
