+++ src/utils/translations/aboutUs.ts
@@ -1,0 +1,39 @@
+export const about_us = {
+  about_us: {
+    en: "About Us",
+    se: "Om Oss",
+  },
+  first_description: {
+    en: "You will witness God shaping the divided seas as we bring together our own divided forms, ultimately replenishing them by the end of this process.",
+    se: "Du kommer att bevittna hur Gud formar de delade haven när vi sammanför våra egna delade former, och till slut fyller på dem i slutet av denna process.",
+  },
+  fast_service: {
+    en: "Fast Service",
+    se: "Snabb Service",
+  },
+  fresh_food: {
+    en: "Fresh Food",
+    se: "Färsk mat",
+  },
+  fast_support: {
+    en: "Fast Support",
+    se: "Snabb Support",
+  },
+  second_title: {
+    en: "We Speak the Language of Foods",
+    se: "Vi talar matens språk",
+  },
+  second_description: {
+    en: "Welcome to BUDDHA, where culinary artistry meets a warm and inviting atmosphere. Nestled in the heart of Stockholm, our restaurant is a gastronomic haven that promises an unforgettable dining experience. From the moment you step through our doors, you'll be transported into a world of exquisite flavors, impeccable service, and a welcoming ambiance that feels like home. Whether you're seeking a romantic dinner for two, a memorable celebration with friends and family, or simply a delightful meal to savor, we invite you to join us and indulge in a culinary journey that celebrates the finest in food and hospitality.",
+    se: "Välkommen till BUDDHA, där kulinariskt konstnärskap möter en varm och inbjudande atmosfär. Inbäddat i hjärtat av Stockholm är vår restaurang en gastronomisk fristad som utlovar en oförglömlig matupplevelse. Från det ögonblick du kliver in genom våra dörrar kommer du att transporteras in i en värld av utsökta smaker, oklanderlig service och en välkomnande atmosfär som känns som hemma. Oavsett om du letar efter en romantisk middag för två, en minnesvärd fest med vänner och familj, eller bara en härlig måltid att njuta av, bjuder vi in ​​dig att gå med oss ​​och njuta av en kulinarisk resa som firar det finaste inom mat och gästfrihet.",
+  },
+  featured_food: {
+    en: "Featured Food",
+    se: "Utvald Mat",
+  },
+  featured_food_info: {
+    en: "Fresh taste and great price",
+    se: "Fräsch smak och bra pris",
+  },
+};
+
