+++ src/utils/translations/catering.ts
@@ -1,0 +1,286 @@
+export const catering = {
+  description: {
+    en: "Indulge in a unique dining experience at our separate serving room, connected to our restaurant. Perfect for birthdays, meetings, and private parties. Choose from our extensive menu or rent our event venue for your next gathering. Call us today for more details and to book your special event at our restaurant.",
+    se: "Njut av en unik matupplevelse i vårt separata serveringsrum, anslutet till vår restaurang. Perfekt för födelsedagar, möten och privata fester. Välj från vår omfattande meny eller hyr vår evenemangslokal för din nästa sammankomst. Ring oss idag för mer information och för att boka ditt speciella evenemang på vårt restaurang",
+  },
+
+  chef_info: {
+    title: {
+      en: "Food should be",
+      se: "Mat ska vara",
+    },
+    text_one: {
+      en: "Abid Khan began his cooking journey in 1991 in Sweden. Today, together with his wife, he runs the Cumin Club restaurants on Södermannagatan and Luntmakargatan in Stockholm. He cooks food from the heart, food that his guests enjoy and that children enjoy.",
+      se: "Abid Khan började sin matlagningsresa 1991 i Sverige. Idag driver han, tillsammans med sin fru, restaurangerna Cumin Club på Södermannagatan och på Luntmakargatan i Stockholm. Han lagar mat från hjärtat, mat som hans gäster tycker om och som barn tycker om.",
+    },
+    text_two: {
+      en: "Abid Khan is passionate about his cooking and his own spice mixes. He is careful to make everything from scratch with natural ingredients and spices 'I don't trust ready-made powders,' he says. Once or twice a week he stands and mixes and mortars his various seasonings for Tikka Masala, Korma and the other classic spice mixes.",
+      se: "Abid Khan brinner för sin matlagning och sina egna kryddblandningar. Han är noga med att göra allt från grunden med naturliga råvaror och kryddor “Jag litar inte på färdiga pulver”, säger han. En till två gånger per vecka står han och mixar och mortlar sina olika smaksättningar till Tikka Masala, Korma och de andra klassiska kryddblandningarna.",
+    },
+    text_three: {
+      en: "Abid also never uses gluten or lactose (except yogurt) in its food. 'The taste is not in wheat or cream', says Abid. 'I have never used these ingredients in my kitchen,' he continues. It suits today's society very well as Abid has noticed that many of his guests are allergic or want to avoid these particular substances.",
+      se: "Abid använder heller aldrig gluten eller laktos (förutom yoghurt) i sin mat. “Smaken sitter inte i vete eller grädde”, menar Abid. “Jag har aldrig använt dessa råvaror i mitt kök”, fortsätter han. Det passar dagens samhälle väldigt bra då Abid märkt att många av hans gäster är allergiker eller vill undvika just dessa ämnen.",
+    },
+  },
+
+  category: [
+    {
+      name: {
+        en: "Starters & Salads",
+        se: "Förrätt och Sallad",
+      },
+      source: "/starters.png",
+      /*description: {
+        en: "Choose any 2",
+        se: "Välj valfri 2",
+      },*/
+      items: [
+        /*{
+          name: {
+            en: "Onion Bhaji",
+            se: "Onion Bhaji",
+          },
+          source: "/onion-bhaji.jpg",
+          description: {
+            en: "Fried onions breaded in chickpea flour.",
+            se: "Stekt lök panerad i kikärtsmjöl.",
+          },
+        },
+        {
+          name: {
+            en: "Papadam",
+            se: "Papadam",
+          },
+          source: "/papadam.jpg",
+
+          description: {
+            en: "Crispy lentil bread with mango chutney.",
+            se: "Krispigt linsbröd med mango chutney.",
+          },
+        },
+        {
+          name: {
+            en: "Shekh Kebab",
+            se: "Shekh Kebab",
+          },
+          source: "/shekh-kebab.jpg",
+          description: {
+            en: "Lamb mince mixed with finely chopped onion, chilli, garlic, coriander and ginger, rolled on a skewer and grilled in a clay oven.",
+            se: "Lammmynta blandad med finhakad lök, chili, vitlök, koriander och ingefära, rullad på spett och grillad i leraugn.",
+          },
+        },
+        {
+          name: {
+            en: "Samosa Chaat",
+            se: "Samosa Chaat",
+          },
+          source: "/samosa-chaat.jpg",
+
+          description: {
+            en: "Samosa Chaat - Homemade vegetarian samosa, topped with raita, cucumber, tomato, onion, coriander tamarind sauce and mint chutney.",
+            se: "Hemlagad vegetarisk samosa, toppad med raita, gurka, tomat, lök, koriander tamarindsås och mynta cutney.",
+          },
+        },
+        {
+          name: {
+            en: "Panir Salad",
+            se: "Panir Salad",
+          },
+          source: "/paneer-salad.jpeg",
+          description: {
+            en: "Mix salad, homemade Indian cheese, olive oil, lime juice.",
+            se: "Mix sallad, hemlagad indisk ost, olivolja, limesaft.",
+          },
+        }, */
+      ],
+    },
+    {
+      name: {
+        en: "Small Buffet",
+        se: "Mindre Buffé",
+      },
+      source: "/buffet.png",
+      items: [
+        /*{
+          name: {
+            en: "Onion Bhaji",
+            se: "Onion Bhaji",
+          },
+          source: "/onion-bhaji.jpg",
+
+          description: {
+            en: "Fried onions breaded in chickpea flour.",
+            se: "Stekt lök panerad i kikärtsmjöl.",
+          },
+        },
+        {
+          name: {
+            en: "Papadam",
+            se: "Papadam",
+          },
+          source: "/papadam.jpg",
+          description: {
+            en: "Crispy lentil bread with mango chutney.",
+            se: "Krispigt linsbröd med mango chutney.",
+          },
+        },
+        {
+          name: {
+            en: "Chicken Tikka Masala",
+            se: "Kyckling Tikka Masala",
+          },
+          source: "/chicken-tikka.jpg",
+        },
+        {
+          name: {
+            en: "Mix Vegetarian Masala (Vegan)",
+            se: "Mix Vegetarisk Masala (Vegon)",
+          },
+          source: "/mix-veg.jpeg",
+
+          description: {
+            en: "Mixed vegetables in a tomato-based vegetarian vegan sauce.",
+            se: "Blandade grönsaker i en tomatbaserad vegetarisk vegon sås.",
+          },
+        },
+        {
+          name: {
+            en: "Naan Bread",
+            se: "Naanbröd",
+          },
+          source: "/naan.jpg",
+        }, */
+      ],
+      price: "   ",
+      delivery: {
+        en: "   ",
+        se: "   ",
+      },
+    },
+    {
+      name: {
+        en: "Main Course ",
+        se: "Huvudrets ",
+      },
+      source: "/main-course.png",
+      description: {
+        en: " ",
+        se: " ",
+      },
+      items: [
+        /*{
+          name: {
+            en: "Mixed Chicken Grill",
+            se: "Mixed Kyckling Grill",
+          },
+          source: "/mixed-grill.jpg",
+          description: {
+            en: "Chicken fillet marinated in tandoori, garlic and green chili, grilled in a clay oven and served with rice, fresh vegetables and sauce",
+            se: "Kycklingfilé marinerad i tandoori, vitlök och grön chili, grillad i leraugn och serveras med ris, färska grönsaker och sås",
+          },
+        },
+        {
+          name: {
+            en: "Vindaloo Steak",
+            se: "Vindaloo Steak",
+          },
+          source: "/vindaloo-steak.jpg",
+
+          description: {
+            en: "A very hearty South Indian dish served with rice",
+            se: "En mycket stark maträtt från södra Indien serverad medris",
+          },
+        },
+        {
+          name: {
+            en: "Chicken Tikka Butter Masala",
+            se: "Kyckling Tikka Butter Masala",
+          },
+          source: "/chicken-butter.jpeg",
+
+          description: {
+            en: "Grilled chicken breast fillet cooked with butter fried spices served with rice",
+            se: "Grillad kycklingbröstfilé tillagad med smörstekt kryddor serverad med ris",
+          },
+        },
+        {
+          name: {
+            en: "Biryani",
+            se: "Biryani",
+          },
+          source: "/biryani.jpeg",
+          description: {
+            en: "Indian saffron risotto with oriental spices. Choose between steak and vegetarian. Served with masala sauce and vegetables",
+            se: "Indisk saffranrisotto med orientaliska kryddor. Välj mellan biff och vegetarian. Serveras med masalasås och grönsaker",
+          },
+        },
+        {
+          name: {
+            en: "Palak Panir",
+            se: "Palak Panir",
+          },
+          source: "/palak-paneer.jpg",
+          description: {
+            en: "Homemade cream cheese cooked with spinach served with rice.",
+            se: "Hemlagad gräddost tillagad med spenat serverad med ris.",
+          },
+        },
+        {
+          name: {
+            en: "Vegan Kofte",
+            se: "Vegan Kofte",
+          },
+          source: "/veg-kofte.jpg",
+          description: {
+            en: "Vegetable buns in a spinach sauce (gluten and dairy free)",
+            se: "Vegetabiliska bullar i en spenatsås (gluten- och mjölkfri)",
+          },
+        }, */
+      ],
+    },
+    {
+      name: {
+        se: "Dessert",
+        en: "Dessert",
+      },
+      source: "/dessert.png",
+      description: {
+        en: " ",
+        se: " ",
+      },
+      items: [
+        /*{
+          name: {
+            en: "",
+            se: "",
+          },
+          source: "/gulab-jamun.jpg",
+
+          description: {
+            en: "",
+            se: "",
+          },
+        },
+        {
+          name: {
+            en: "",
+            se: "",
+          },
+          source: "/fruit-vanilla.jpg",
+        },
+        {
+          name: {
+            en: "",
+            se: "",
+          },
+          source: "/gajar-halwa.jpg",
+          description: {
+            en: "",
+            se: "",
+          },
+        },*/
+      ],
+    },
+  ],
+};
+
