+++ src/utils/translations/contact.ts
@@ -1,0 +1,55 @@
+export const contact = {
+  title: {
+    en: "Contact",
+    se: "Kontakt",
+  },
+  address_one: {
+    en: "Årstavägen 39, 120 52 Årsta",
+    se: "Årstavägen 39, 120 52 Årsta",
+  },
+  address_two: {
+    en: "Ynglingagatan 9, 113 47 Stockholm",
+    se: "Ynglingagatan 9, 113 47 Stockholm",
+  },
+  opening_times: {
+    en: "Opening times for this location:",
+    se: "Öppettider för denna plats:",
+  },
+  weekdays: {
+    en: "Mon - Fri : 11:00 - 21:00",
+    se: "Mån - Fre : 11:00 - 21:00",
+  },
+  weekends: {
+    en: "Sat - Sun : 13:00 - 21:00",
+    se: "Lör - Sön : 13:00 - 21:00",
+  },
+  order_call: {
+    en: "You can order by calling us on:",
+    se: "Du kan beställa genom att ringa oss på:",
+  },
+  open_label: {
+    en: "Open",
+    se: "Öppet",
+  },
+  coming_soon_default: {
+    en: "Coming soon",
+    se: "Kommer snart",
+  },
+  coming_soon_description: {
+    en: "We're preparing something special for this neighborhood.",
+    se: "Vi förbereder något speciellt för det här området.",
+  },
+  coming_soon_follow: {
+    en: "Follow us for updates and be the first to know when BUDDHA Haga opens its doors.",
+    se: "Följ oss för uppdateringar och var först med att få veta när BUDDHA Haga öppnar sina dörrar.",
+  },
+  phone_one: {
+    en: "08-684 271 90",
+    se: "08-684 271 90",
+  },
+  phone_two: {
+    en: "0760-35 37 99",
+    se: "0760-35 37 99",
+  },
+};
+
