+++ src/utils/translations/drinks.ts
@@ -1,0 +1,431 @@
+export const drinks = {
+  bottle_beer: {
+    name: {
+      en: "Bottle Beer cl Dragon Beer",
+      se: "Flask Öl cl Fatöl",
+    },
+    items: [
+      {
+        name: {
+          en: "Bangla",
+          se: "Bangla",
+        },
+      },
+      {
+        name: {
+          en: "Indian Beer; Light Stock Little",
+          se: "Indisk Öl; Ljus Lager Lite",
+        },
+        price: "33, 55",
+      },
+      {
+        name: {
+          en: "Fuller",
+          se: "Fylligare",
+        },
+        price: "66 95",
+      },
+      {
+        name: {
+          en: "Cobra",
+          se: "Cobra",
+        },
+        price: "33, 55",
+      },
+      {
+        name: {
+          en: "Light Lager Indian Beer",
+          se: "Ljus Lager Indish Öl",
+        },
+        price: "66, 95",
+      },
+      {
+        name: {
+          en: "Kingfisher",
+          se: "Kingfisher",
+        },
+      },
+      {
+        name: {
+          en: "Indian Light Foam",
+          se: "Indish Ljust Skum",
+        },
+        price: "33, 55",
+      },
+      {
+        name: {
+          en: "Bombay Gold",
+          se: "Bombay Gold",
+        },
+        price: "66, 86",
+      },
+      {
+        name: {
+          en: "Mariestads",
+          se: "Mariestads",
+        },
+        price: "50, 59",
+      },
+      {
+        name: {
+          en: "Cider (Pear, Mango, Apple),",
+          se: "Cider (Päron, Mango, Äppel),",
+        },
+        price: "33, 55",
+      },
+      {
+        name: {
+          en: "Folk Beer",
+          se: "Folköl",
+        },
+        price: "33, 49",
+      },
+      {
+        name: {
+          en: "Non-Alcoholic Beer",
+          se: "Alkoholfritt Öl",
+        },
+        price: "33, 39",
+      },
+      {
+        name: {
+          en: "Draft Beer",
+          se: "Fatöl",
+        },
+      },
+      {
+        name: {
+          en: "Spendrups Lager 5.3%",
+          se: "Spendrups Lager 5,3%",
+        },
+        price: "40, 49",
+      },
+      {
+        name: {
+          en: "Sitting Bulldog IPA 5.8%",
+          se: "Sitting Bulldog IPA 5,8%",
+        },
+        price: "40, 49",
+      },
+    ],
+  },
+  juice: {
+    name: {
+      en: "Lassi/Juice",
+      se: "Läsk/Juice",
+    },
+    items: [
+      {
+        name: {
+          en: "Cola, Fanta, Loka, Sprite, Zero",
+          se: "Cola, Fanta, Loka, Sprite, Zero",
+        },
+        price: "29",
+      },
+      {
+        name: {
+          en: "Mango Juice",
+          se: "Mango Juice",
+        },
+        price: "39",
+      },
+      {
+        name: {
+          en: "Mango Lassi",
+          se: "Mango Lassi",
+        },
+        price: "49",
+      },
+    ],
+  },
+  hot_beverage: {
+    name: {
+      en: "Hot Beverages",
+      se: "Varma Drycker",
+    },
+    items: [
+      {
+        name: {
+          en: "Coffee",
+          se: "Kaffe",
+        },
+        price: "25",
+      },
+      {
+        name: {
+          en: "Irish Coffee 4CL",
+          se: "Irish Kaffe 4CL",
+        },
+        price: "86",
+      },
+      {
+        name: {
+          en: "Indian Chai",
+          se: "Indisk Chai",
+        },
+        price: "25",
+      },
+    ],
+  },
+  drinks: {
+    name: {
+      en: "Drinks",
+      se: "Drinker",
+    },
+    description: {
+      en: "4cl-109 kr, 6cl-129 kr",
+      se: "4cl-109 kr, 6cl-129 kr",
+    },
+    items: [
+      {
+        name: {
+          en: "Gin Tonic",
+          se: "Gin Tonic",
+        },
+        description: {
+          en: "(Gin, Tonic, Citron)",
+          se: "(Gin, Tonic, Citron)",
+        },
+      },
+      {
+        name: {
+          en: "Vodka Lime",
+          se: "Vodka Lime",
+        },
+        description: {
+          en: "(Bodka, Krossad IS, Lime Fruit)",
+          se: "(Bodka, Krossad IS, Lime Fruit)",
+        },
+      },
+
+      {
+        name: {
+          en: "Whisky Sour",
+          se: "Whisky Sour",
+        },
+        description: {
+          en: "(Whiskey, Citronsaft)",
+          se: "(Whiskey, Citronsaft)",
+        },
+      },
+      {
+        name: {
+          en: "Cava",
+          se: "Cava",
+        },
+        description: {
+          en: "(Whiskey, Citronsaft)",
+          se: "(Whiskey, Citronsaft)",
+        },
+      },
+      {
+        name: {
+          en: "Glas",
+          se: "Glas",
+        },
+        price: "69",
+      },
+
+      {
+        name: {
+          en: "Flaska",
+          se: "Flaska",
+        },
+        price: "276",
+      },
+      {
+        name: {
+          en: "House Wine Red",
+          se: "Husets Vin Rött",
+        },
+      },
+      {
+        name: {
+          en: "Torre Del Falasco Corvina IGT-Red (Italy)",
+          se: "Torre Del Falasco Corvina IGT-Rött",
+        },
+        description: {
+          en: "Medium-bodied wine with notes of blackberries, morels, dark chocolate, cherry pits and brisk acidity and fruity, long aftertaste.",
+          se: "Medelfylligt vin med toner av björnbär, moreller, mörk choklad, körsbärskärnor samt pigg syra och fruktig, lång eftersmak.",
+        },
+        price: "75, 295",
+      },
+      {
+        name: {
+          en: "House Wine White",
+          se: "Husets Vin Vitt",
+        },
+        description: {
+          en: "Medium-bodied wine with fresh acidity, good concentration as well, mineral rich, long, spicy taste and aftertaste",
+          se: "Medelfylligt vin med fräsch syra, bra koncentration samt Mineralrich, lång, kryddig smak och eftersmak.",
+        },
+        price: "75, 295",
+      },
+    ],
+  },
+  red_wine: {
+    name: {
+      en: "Red Wine",
+      se: "Rött Vin",
+    },
+    items: [
+      {
+        name: {
+          en: "Vina Maipo",
+          se: "Vina Maipo",
+        },
+        sub: [
+          {
+            name: {
+              en: "Cabernet Sauvignon(Chille)",
+              se: "Cabernet Sauvignon(Chille)",
+            },
+            description: {
+              en: "This bright ruby red has an attractive aroma of fresh red fruits,with notes of spices and tobacco. The palate is juicy, with harmonic tannins and good aftertaste.",
+              se: "Denna ljusa rubinröda har en attraktiv doft av färska röda frukter, med toner av kryddor och tobak. Smaken är saftig, med harmoniska tanniner och god eftersmak.",
+            },
+            price: "235",
+          },
+        ],
+      },
+      {
+        name: {
+          en: "Costa",
+          se: "Costa",
+        },
+        sub: [
+          {
+            name: {
+              en: "Salento Negroamaro(Italia)",
+              se: "Salento Negroamaro(Italia)",
+            },
+            price: "65, 260",
+          },
+          {
+            name: {
+              en: "Grand Vin De Bordeaux (France) Flaska",
+              se: "Grand Vin De Bordeaux (France) Flaska",
+            },
+            price: "395",
+          },
+          {
+            name: {
+              en: "Luma Rosso I Sicilia (Italy)",
+              se: "Luma Rosso I Sicilia (Italy)",
+            },
+            price: "265",
+          },
+          {
+            name: {
+              en: "Roma Ross (Italian)",
+              se: "Roma Ross (Italian)",
+            },
+            price: "325",
+          },
+          {
+            name: {
+              en: "Dark Horse Pinot Noir (USA)",
+              se: "Dark Horse Pinot Noir (USA)",
+            },
+            description: {
+              en: "An elegant Pinot Noir with notes of cherry and a layered, flavorful finish.",
+              se: "En elegant Pinot Noir med noter av körsbär och en skiktad, smakrik finish.",
+            },
+            price: "350",
+          },
+        ],
+      },
+      {
+        name: {
+          en: "Vitt Wine",
+          se: "Vitt Vin",
+        },
+        sub: [
+          {
+            name: {
+              en: "LA Merdionale Ordinary Sauvignon",
+              se: "LA Merdionale Vanlig Sauvignon",
+            },
+            description: {
+              en: "Wine with aromas of pink grapefruit and lemons.",
+              se: "Vin med aromer av rosa grapefrukt och citroner.",
+            },
+            price: "65, 265",
+          },
+          {
+            name: {
+              en: "Bianco Salento (Italian)",
+              se: "Bianco Salento (Italien)",
+            },
+            description: {
+              en: "Medium-bodied wine with pleasant warmth, freshness, fruity taste of peaches and melon. Good balance and with a small, fresh aftertaste.",
+              se: "Medelfylligt vin med trevlig värme, fräschör, fruktig smak av persikor och melon. Bra balans och med liten, frisk eftersmak.",
+            },
+          },
+          {
+            name: {
+              en: "Dark Horse Pinot Grigio (USA)",
+              se: "Dark Horse Pinot Grigio (USA)",
+            },
+            description: {
+              en: "A dry Pinot Grigio with notes of stonefruit and a balanced, crisp finish.",
+              se: "En torr Pinot Grigio med toner av stenfrukt och en balanserad, krispig finish.",
+            },
+          },
+        ],
+      },
+      {
+        name: {
+          en: "Rose Wine",
+          se: "Rose Vin",
+        },
+        sub: [
+          {
+            name: {
+              en: "Vina Maipo",
+              se: "Vina Maipo",
+            },
+            description: {
+              en: "Bright delicate pink color wine with fruity aromas, especially wild berries and pomegranate. In palate has good acidity, with mineral and fresh red fruits flavors.",
+              se: "Ljust, delikat rosa färg vin med fruktiga aromer, särskilt vilda bär och granatäpple. I gommen har bra syra, med mineral och färska röda frukter smaker.",
+            },
+            price: "235",
+          },
+        ],
+      },
+      {
+        name: {
+          en: "Alcohol Free",
+          se: "Alkoholfritt",
+        },
+        sub: [
+          {
+            name: {
+              en: "Nature Red",
+              se: "Nature Red",
+            },
+            price: "215",
+          },
+          {
+            name: {
+              en: "Torre Spanien",
+              se: "Torre Spanien",
+            },
+            description: {
+              en: "Light wine with hints of pomegranate, red currants and vanilla.",
+              se: "Lätt vin med inslag av granatäpple, röda vinbär och vanilj.",
+            },
+          },
+          {
+            name: {
+              en: "Chapel Hill Sparkling",
+              se: "Chapel Hill Sparkling",
+            },
+            price: "215",
+          },
+        ],
+      },
+    ],
+  },
+};
+
