+++ src/utils/translations/footerContent.ts
@@ -1,0 +1,43 @@
+export const footerContent = {
+  tagline: {
+    en: "Authentic Himalayan flavors in the heart of Stockholm.",
+    se: "Äkta himalayiska smaker i Stockholms hjärta.",
+  },
+  ctaLabel: {
+    en: "Book a Table",
+    se: "Boka bord",
+  },
+  ctaDescription: {
+    en: "Reserve your seat and let us prepare a feast for you and your company.",
+    se: "Boka ditt bord så förbereder vi en fest för dig och ditt sällskap.",
+  },
+  exploreTitle: {
+    en: "Explore",
+    se: "Utforska",
+  },
+  contactTitle: {
+    en: "Visit & Contact",
+    se: "Besök & kontakt",
+  },
+  visitUsTitle: {
+    en: "Visit us",
+    se: "Besök oss",
+  },
+  callUsTitle: {
+    en: "Call us",
+    se: "Ring oss",
+  },
+  hoursTitle: {
+    en: "Opening hours",
+    se: "Öppettider",
+  },
+  deliveryTitle: {
+    en: "Delivery partners",
+    se: "Leveranspartners",
+  },
+  rights: {
+    en: "© 2024 BUDDHA Nepal. All rights reserved.",
+    se: "© 2024 BUDDHA Nepal. Alla rättigheter förbehållna.",
+  },
+};
+
