+++ src/utils/translations/index.ts
@@ -1,0 +1,13 @@
+export type { Language } from "./types";
+export { homepage } from "./homepage";
+export { menu } from "./menu";
+export type { MenuData, MenuSectionKey } from "./menu";
+export { MENU_ICONS } from "./menuIcons";
+export { menuByLocation } from "./menuByLocation";
+export { drinks } from "./drinks";
+export { catering } from "./catering";
+export { about_us } from "./aboutUs";
+export { timings } from "./timings";
+export { contact } from "./contact";
+export { footerContent } from "./footerContent";
+export { messages } from "./messages";
