+++ src/utils/translations/menuIcons.ts
@@ -1,0 +1,13 @@
+import type { MenuSectionKey } from "./menu";
+
+export const MENU_ICONS: Record<MenuSectionKey, string> = {
+  appetizer: "/appetizer.svg",
+  favourites: "/favourites.svg",
+  thali: "/thali.svg",
+  veg: "/veg.svg",
+  clay_oven: "/clay_oven.svg",
+  biryani_dishes: "/mughal.svg",
+  momos: "/momos.png",
+  bread: "/naan.svg",
+  accessories: "/accessories.svg",
+};
