+++ src/utils/translations/messages.ts
@@ -1,0 +1,18 @@
+export const messages = {
+  information: {
+    lunch: {
+      title: {
+        en: "Dear customers!",
+        se: "Kära kunder!",
+      },
+      message: {
+        en: "We will start serving Lunch Buffet from August 19, 2024.",
+        se: "Vi börjar servera Lunch Buffé från och med 19 Augusti, 2024.",
+      },
+      footer: {
+        en: "You are warmly welcome.",
+        se: "Ni är varmt välkomna.",
+      },
+    },
+  },
+};
