+++ src/utils/translations/timings.ts
@@ -1,0 +1,15 @@
+export const timings = {
+  opening_hours: {
+    en: "OPENING HOURS",
+    se: "ÖPPETTIDER",
+  },
+  weekdays: {
+    en: "Mon - Fri 11:00 - 21:00",
+    se: "Mån - Fre 11:00 - 21:00",
+  },
+  weekends: {
+    en: "Sat - Sun 13:00 - 21:00",
+    se: "Lör - Sön 13:00 - 21:00",
+  },
+};
+
