+++ src/utils/translations/types.ts
@@ -1,0 +1,4 @@
+export type Language = {
+  en: string;
+  se: string;
+};
